refactor: move mockSCM behind test build tag
