+++ pkg/process/metadata/parser/scm_reader_windows.go
@@ -8,10 +8,6 @@
 package parser
 
 import (
-	"testing"
-
-	"github.com/stretchr/testify/mock"
-
 	"github.com/DataDog/datadog-agent/pkg/util/winutil"
 )
 
@@ -46,28 +42,4 @@ func (s *scmReader) getServiceInfo(pid uint64) (*WindowsServiceInfo, error) {
 		ServiceName: monitorServiceInfo.ServiceName,
 		DisplayName: monitorServiceInfo.DisplayName,
 	}, nil
-}
-
-var _ mockableSCM = (*mockSCM)(nil)
-
-type mockSCM struct {
-	mock.Mock
-}
-
-func newSCMReaderWithMock(t *testing.T) (*scmReader, *mockSCM) {
-	m := &mockSCM{}
-	m.Test(t)
-	t.Cleanup(func() { m.AssertExpectations(t) })
-	return &scmReader{
-		scmMonitor: m,
-	}, m
-}
-
-func (m *mockSCM) GetServiceInfo(pid uint64) (*winutil.ServiceInfo, error) {
-	args := m.Called(pid)
-
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*winutil.ServiceInfo), args.Error(1)
 }
